Make ExtVars independent of native tree handlers. It is now part of the agent's tree manager component.
These changes leave to the user the possibility to build his tree handler out of the agent source directory.
Now the part responsive to send back informations to the Lua VM is totally hidden from the user.
He just needs to export a specific API in order to make his tree handler compatible with the agent.It fixes bug #404901.